refactor: 회원/비회원 서비스 분리 및 주문 서비스 리팩토링 fix: 로깅 관련 설정 수정
